+++ src/test/java/org/tinyconfiguration/imp/basic/io/CsvTest.java
@@ -2,13 +2,12 @@ package org.tinyconfiguration.imp.basic.io;
 
 import org.junit.jupiter.api.Test;
 import org.tinyconfiguration.imp.basic.Configuration;
-import org.tinyconfiguration.imp.basic.ConfigurationIO;
 import org.tinyconfiguration.imp.basic.Property;
 
 import java.util.concurrent.Future;
 
 import static org.junit.jupiter.api.Assertions.*;
-import static org.tinyconfiguration.abc.utils.ExportType.CSV;
+import static org.tinyconfiguration.abc.utils.FormatType.CSV;
 
 public class CsvTest {
 
@@ -83,13 +82,13 @@ public class CsvTest {
     void readCSV() {
 
         // Does it exists?
-        if (!ConfigurationIO.as(CSV).exist(instance)) {
+        if (!instance.exist()) {
 
             // If so, let's executing the writing task, then execute it with get()
-            assertDoesNotThrow(() -> ConfigurationIO.as(CSV).write(instance));
+            assertDoesNotThrow(() -> instance.write(CSV));
 
             // Now, it should exists
-            assertTrue(ConfigurationIO.as(CSV).exist(instance));
+            assertTrue(instance.exist());
 
         }
 
@@ -105,7 +104,7 @@ public class CsvTest {
         assertArrayEquals(new int[]{10, 15}, instance.get("special-digits").getValue().asIntArray());
 
         // Now, reading the configuration instance
-        assertDoesNotThrow(() -> ConfigurationIO.as(CSV).read(this.instance));
+        assertDoesNotThrow(() -> instance.read(CSV));
 
         assertEquals("root", instance.get("user").getValue().asString());
         assertEquals("toor", instance.get("password").getValue().asString());
@@ -121,7 +120,7 @@ public class CsvTest {
 
             int x = 0;
 
-            Future<Void> task = ConfigurationIO.as(CSV).readAsync(this.instance);
+            Future<Void> task = instance.readAsync(CSV);
 
             while (!task.isDone()) {
                 // Do something
@@ -131,7 +130,7 @@ public class CsvTest {
             assertTrue(x >= 0);
         });
 
-        assertTrue(ConfigurationIO.as(CSV).exist(this.instance));
+        assertTrue(instance.exist());
 
     }
 
@@ -142,7 +141,7 @@ public class CsvTest {
 
             int x = 0;
 
-            Future<Void> task = ConfigurationIO.as(CSV).writeAsync(this.instance);
+            Future<Void> task = instance.writeAsync(CSV);
 
             while (!task.isDone()) {
                 ++x;
@@ -152,34 +151,34 @@ public class CsvTest {
 
         });
 
-        assertTrue(ConfigurationIO.as(CSV).exist(instance));
+        assertTrue(instance.exist());
 
     }
 
     @Test
     void writeCSV() {
-        assertDoesNotThrow(() -> ConfigurationIO.as(CSV).write(this.instance));
-        assertTrue(ConfigurationIO.as(CSV).exist(instance));
+        assertDoesNotThrow(() -> instance.write(CSV));
+        assertTrue(instance.exist());
     }
 
     @Test
     void deleteCSV() {
 
         // Does it exists?
-        if (!ConfigurationIO.as(CSV).exist(instance)) {
+        if (!instance.exist()) {
 
-            // If so, let's executing the writing task, then execute it with get()
-            assertDoesNotThrow(() -> ConfigurationIO.as(CSV).write(instance));
+            // If so, let's executing the writing task
+            assertDoesNotThrow(() -> instance.write(CSV));
 
             // Now, it should exists
-            assertTrue(ConfigurationIO.as(CSV).exist(instance));
+            assertTrue(instance.exist());
 
         }
 
         // Executing deleting task, then execute it
-        assertDoesNotThrow(() -> ConfigurationIO.as(CSV).delete(instance));
+        assertDoesNotThrow(instance::delete);
         // Asserting does not exists any more
-        assertFalse(ConfigurationIO.as(CSV).exist(instance));
+        assertFalse(instance.exist());
 
     }
 
@@ -187,19 +186,19 @@ public class CsvTest {
     void deleteAsyncCSV() {
 
         // Does it exists?
-        if (!ConfigurationIO.as(CSV).exist(instance)) {
+        if (!instance.exist()) {
 
             // If so, let's obtain an a-sync writing task, then execute it with get()
-            assertDoesNotThrow(() -> ConfigurationIO.as(CSV).writeAsync(instance).get());
+            assertDoesNotThrow(() -> instance.writeAsync(CSV));
 
             // Now, it should exists
-            assertTrue(ConfigurationIO.as(CSV).exist(instance));
+            assertTrue(instance.exist());
 
         }
         // Obtaining deleting task, then execute it
-        assertDoesNotThrow(() -> ConfigurationIO.as(CSV).deleteAsync(instance).get());
+        assertDoesNotThrow(() -> instance.deleteAsync().get());
         // Asserting does not exists any more
-        assertFalse(ConfigurationIO.as(CSV).exist(instance));
+        assertFalse(instance.exist());
 
     }