+++ src/test/java/org/tinyconfiguration/imp/basic/io/JsonTest.java
@@ -2,13 +2,12 @@ package org.tinyconfiguration.imp.basic.io;
 
 import org.junit.jupiter.api.Test;
 import org.tinyconfiguration.imp.basic.Configuration;
-import org.tinyconfiguration.imp.basic.ConfigurationIO;
 import org.tinyconfiguration.imp.basic.Property;
 
 import java.util.concurrent.Future;
 
 import static org.junit.jupiter.api.Assertions.*;
-import static org.tinyconfiguration.abc.utils.ExportType.JSON;
+import static org.tinyconfiguration.abc.utils.FormatType.JSON;
 
 class JsonTest {
 
@@ -83,9 +82,9 @@ class JsonTest {
     @Test
     void writeJSON() {
 
-        assertDoesNotThrow(() -> ConfigurationIO.as(JSON).write(this.instance));
+        assertDoesNotThrow(() -> instance.write(JSON));
 
-        assertTrue(ConfigurationIO.as(JSON).exist(instance));
+        assertTrue(instance.exist());
 
     }
 
@@ -96,7 +95,7 @@ class JsonTest {
 
             int x = 0;
 
-            Future<Void> task = ConfigurationIO.as(JSON).writeAsync(this.instance);
+            Future<Void> task = instance.writeAsync(JSON);
 
             while (!task.isDone()) {
                 ++x;
@@ -106,7 +105,7 @@ class JsonTest {
 
         });
 
-        assertTrue(ConfigurationIO.as(JSON).exist(instance));
+        assertTrue(instance.exist());
 
     }
 
@@ -114,13 +113,13 @@ class JsonTest {
     void readJSON() {
 
         // Does it exists?
-        if (!ConfigurationIO.as(JSON).exist(instance)) {
+        if (!instance.exist()) {
 
             // If so, let's executing the writing task, then execute it with get()
-            assertDoesNotThrow(() -> ConfigurationIO.as(JSON).write(instance));
+            assertDoesNotThrow(() -> instance.write(JSON));
 
             // Now, it should exists
-            assertTrue(ConfigurationIO.as(JSON).exist(instance));
+            assertTrue(instance.exist());
 
         }
 
@@ -136,7 +135,7 @@ class JsonTest {
         assertArrayEquals(new int[]{10, 15}, instance.get("special-digits").getValue().asIntArray());
 
         // Now, reading the configuration instance
-        assertDoesNotThrow(() -> ConfigurationIO.as(JSON).read(this.instance));
+        assertDoesNotThrow(() -> instance.read(JSON));
 
         assertEquals("root", instance.get("user").getValue().asString());
         assertEquals("toor", instance.get("password").getValue().asString());
@@ -152,7 +151,7 @@ class JsonTest {
 
             int x = 0;
 
-            Future<Void> task = ConfigurationIO.as(JSON).readAsync(this.instance);
+            Future<Void> task = instance.readAsync(JSON);
 
             while (!task.isDone()) {
                 // Do something
@@ -162,7 +161,7 @@ class JsonTest {
             assertTrue(x >= 0);
         });
 
-        assertTrue(ConfigurationIO.as(JSON).exist(this.instance));
+        assertTrue(instance.exist());
 
     }
 
@@ -170,20 +169,20 @@ class JsonTest {
     void deleteJSON() {
 
         // Does it exists?
-        if (!ConfigurationIO.as(JSON).exist(instance)) {
+        if (!instance.exist()) {
 
             // If so, let's executing the writing task, then execute it with get()
-            assertDoesNotThrow(() -> ConfigurationIO.as(JSON).write(instance));
+            assertDoesNotThrow(() -> instance.write(JSON));
 
             // Now, it should exists
-            assertTrue(ConfigurationIO.as(JSON).exist(instance));
+            assertTrue(instance.exist());
 
         }
 
         // Executing deleting task, then execute it
-        assertDoesNotThrow(() -> ConfigurationIO.as(JSON).delete(instance));
+        assertDoesNotThrow(() -> instance.deleteAsync().get());
         // Asserting does not exists any more
-        assertFalse(ConfigurationIO.as(JSON).exist(instance));
+        assertFalse(instance.exist());
 
     }
 
@@ -191,19 +190,19 @@ class JsonTest {
     void deleteAsyncJSON() {
 
         // Does it exists?
-        if (!ConfigurationIO.as(JSON).exist(instance)) {
+        if (!instance.exist()) {
 
             // If so, let's obtain an a-sync writing task, then execute it with get()
-            assertDoesNotThrow(() -> ConfigurationIO.as(JSON).writeAsync(instance).get());
+            assertDoesNotThrow(() -> instance.writeAsync(JSON).get());
 
             // Now, it should exists
-            assertTrue(ConfigurationIO.as(JSON).exist(instance));
+            assertTrue(instance.exist());
 
         }
         // Obtaining deleting task, then execute it
-        assertDoesNotThrow(() -> ConfigurationIO.as(JSON).deleteAsync(instance).get());
+        assertDoesNotThrow(() -> instance.deleteAsync().get());
         // Asserting does not exists any more
-        assertFalse(ConfigurationIO.as(JSON).exist(instance));
+        assertFalse(instance.exist());
 
     }
 }