+++ src/test/java/org/tinyconfiguration/imp/basic/io/XmlTest.java
@@ -2,13 +2,12 @@ package org.tinyconfiguration.imp.basic.io;
 
 import org.junit.jupiter.api.Test;
 import org.tinyconfiguration.imp.basic.Configuration;
-import org.tinyconfiguration.imp.basic.ConfigurationIO;
 import org.tinyconfiguration.imp.basic.Property;
 
 import java.util.concurrent.Future;
 
 import static org.junit.jupiter.api.Assertions.*;
-import static org.tinyconfiguration.abc.utils.ExportType.XML;
+import static org.tinyconfiguration.abc.utils.FormatType.XML;
 
 class XmlTest {
 
@@ -83,13 +82,13 @@ class XmlTest {
     @Test
     void readXML() {
         // Does it exists?
-        if (!ConfigurationIO.as(XML).exist(instance)) {
+        if (!instance.exist()) {
 
             // If so, let's executing the writing task, then execute it with get()
-            assertDoesNotThrow(() -> ConfigurationIO.as(XML).write(instance));
+            assertDoesNotThrow(() -> instance.write(XML));
 
             // Now, it should exists
-            assertTrue(ConfigurationIO.as(XML).exist(instance));
+            assertTrue(instance.exist());
 
         }
 
@@ -105,7 +104,7 @@ class XmlTest {
         assertArrayEquals(new int[]{10, 15}, instance.get("special-digits").getValue().asIntArray());
 
         // Now, reading the configuration instance
-        assertDoesNotThrow(() -> ConfigurationIO.as(XML).read(this.instance));
+        assertDoesNotThrow(() -> instance.read(XML));
 
         assertEquals("root", instance.get("user").getValue().asString());
         assertEquals("toor", instance.get("password").getValue().asString());
@@ -121,7 +120,7 @@ class XmlTest {
 
             int x = 0;
 
-            Future<Void> task = ConfigurationIO.as(XML).readAsync(this.instance);
+            Future<Void> task = instance.readAsync(XML);
 
             while (!task.isDone()) {
                 // Do something
@@ -131,16 +130,16 @@ class XmlTest {
             assertTrue(x >= 0);
         });
 
-        assertTrue(ConfigurationIO.as(XML).exist(this.instance));
+        assertTrue(instance.exist());
 
     }
 
     @Test
     void writeXML() {
 
-        assertDoesNotThrow(() -> ConfigurationIO.as(XML).write(this.instance));
+        assertDoesNotThrow(() -> instance.write(XML));
 
-        assertTrue(ConfigurationIO.as(XML).exist(instance));
+        assertTrue(instance.exist());
     }
 
     @Test
@@ -150,7 +149,7 @@ class XmlTest {
 
             int x = 0;
 
-            Future<Void> task = ConfigurationIO.as(XML).writeAsync(this.instance);
+            Future<Void> task = instance.writeAsync(XML);
 
             while (!task.isDone()) {
                 ++x;
@@ -160,7 +159,7 @@ class XmlTest {
 
         });
 
-        assertTrue(ConfigurationIO.as(XML).exist(instance));
+        assertTrue(instance.exist());
 
     }
 
@@ -168,20 +167,20 @@ class XmlTest {
     void deleteXML() {
 
         // Does it exists?
-        if (!ConfigurationIO.as(XML).exist(instance)) {
+        if (!instance.exist()) {
 
             // If so, let's executing the writing task, then execute it with get()
-            assertDoesNotThrow(() -> ConfigurationIO.as(XML).write(instance));
+            assertDoesNotThrow(() -> instance.write(XML));
 
             // Now, it should exists
-            assertTrue(ConfigurationIO.as(XML).exist(instance));
+            assertTrue(instance.exist());
 
         }
 
         // Executing deleting task, then execute it
-        assertDoesNotThrow(() -> ConfigurationIO.as(XML).delete(instance));
+        assertDoesNotThrow(instance::delete);
         // Asserting does not exists any more
-        assertFalse(ConfigurationIO.as(XML).exist(instance));
+        assertFalse(instance.exist());
 
     }
 
@@ -189,19 +188,19 @@ class XmlTest {
     void deleteAsyncXML() {
 
         // Does it exists?
-        if (!ConfigurationIO.as(XML).exist(instance)) {
+        if (!instance.exist()) {
 
             // If so, let's obtain an a-sync writing task, then execute it with get()
-            assertDoesNotThrow(() -> ConfigurationIO.as(XML).writeAsync(instance).get());
+            assertDoesNotThrow(() -> instance.writeAsync(XML).get());
 
             // Now, it should exists
-            assertTrue(ConfigurationIO.as(XML).exist(instance));
+            assertTrue(instance.exist());
 
         }
         // Obtaining deleting task, then execute it
-        assertDoesNotThrow(() -> ConfigurationIO.as(XML).deleteAsync(instance).get());
+        assertDoesNotThrow(() -> instance.deleteAsync().get());
         // Asserting does not exists any more
-        assertFalse(ConfigurationIO.as(XML).exist(instance));
+        assertFalse(instance.exist());
 
     }
 }