+++ src/test/java/org/tinyconfiguration/imp/basic/io/YamlTest.java
@@ -2,13 +2,12 @@ package org.tinyconfiguration.imp.basic.io;
 
 import org.junit.jupiter.api.Test;
 import org.tinyconfiguration.imp.basic.Configuration;
-import org.tinyconfiguration.imp.basic.ConfigurationIO;
 import org.tinyconfiguration.imp.basic.Property;
 
 import java.util.concurrent.Future;
 
 import static org.junit.jupiter.api.Assertions.*;
-import static org.tinyconfiguration.abc.utils.ExportType.YAML;
+import static org.tinyconfiguration.abc.utils.FormatType.YAML;
 
 class YamlTest {
 
@@ -83,13 +82,13 @@ class YamlTest {
     void readYAML() {
 
         // Does it exists?
-        if (!ConfigurationIO.as(YAML).exist(instance)) {
+        if (!instance.exist()) {
 
             // If so, let's executing the writing task, then execute it with get()
-            assertDoesNotThrow(() -> ConfigurationIO.as(YAML).write(instance));
+            assertDoesNotThrow(() -> instance.write(YAML));
 
             // Now, it should exists
-            assertTrue(ConfigurationIO.as(YAML).exist(instance));
+            assertTrue(instance.exist());
 
         }
 
@@ -105,7 +104,7 @@ class YamlTest {
         assertArrayEquals(new int[]{10, 15}, instance.get("special-digits").getValue().asIntArray());
 
         // Now, reading the configuration instance
-        assertDoesNotThrow(() -> ConfigurationIO.as(YAML).read(this.instance));
+        assertDoesNotThrow(() -> instance.read(YAML));
 
         assertEquals("root", instance.get("user").getValue().asString());
         assertEquals("toor", instance.get("password").getValue().asString());
@@ -121,7 +120,7 @@ class YamlTest {
 
             int x = 0;
 
-            Future<Void> task = ConfigurationIO.as(YAML).readAsync(this.instance);
+            Future<Void> task = instance.readAsync(YAML);
 
             while (!task.isDone()) {
                 // Do something
@@ -131,14 +130,14 @@ class YamlTest {
             assertTrue(x >= 0);
         });
 
-        assertTrue(ConfigurationIO.as(YAML).exist(this.instance));
+        assertTrue(instance.exist());
 
     }
 
     @Test
     void writeYAML() {
-        assertDoesNotThrow(() -> ConfigurationIO.as(YAML).write(this.instance));
-        assertTrue(ConfigurationIO.as(YAML).exist(instance));
+        assertDoesNotThrow(() -> instance.write(YAML));
+        assertTrue(instance.exist());
     }
 
     @Test
@@ -148,7 +147,7 @@ class YamlTest {
 
             int x = 0;
 
-            Future<Void> task = ConfigurationIO.as(YAML).writeAsync(this.instance);
+            Future<Void> task = instance.writeAsync(YAML);
 
             while (!task.isDone()) {
                 ++x;
@@ -158,7 +157,7 @@ class YamlTest {
 
         });
 
-        assertTrue(ConfigurationIO.as(YAML).exist(instance));
+        assertTrue(instance.exist());
 
     }
 
@@ -166,20 +165,20 @@ class YamlTest {
     void deleteYAML() {
 
         // Does it exists?
-        if (!ConfigurationIO.as(YAML).exist(instance)) {
+        if (!instance.exist()) {
 
             // If so, let's executing the writing task, then execute it with get()
-            assertDoesNotThrow(() -> ConfigurationIO.as(YAML).write(instance));
+            assertDoesNotThrow(() -> instance.write(YAML));
 
             // Now, it should exists
-            assertTrue(ConfigurationIO.as(YAML).exist(instance));
+            assertTrue(instance.exist());
 
         }
 
         // Executing deleting task, then execute it
-        assertDoesNotThrow(() -> ConfigurationIO.as(YAML).delete(instance));
+        assertDoesNotThrow(instance::delete);
         // Asserting does not exists any more
-        assertFalse(ConfigurationIO.as(YAML).exist(instance));
+        assertFalse(instance.exist());
 
     }
 
@@ -187,19 +186,19 @@ class YamlTest {
     void deleteAsyncYAML() {
 
         // Does it exists?
-        if (!ConfigurationIO.as(YAML).exist(instance)) {
+        if (!instance.exist()) {
 
             // If so, let's obtain an a-sync writing task, then execute it with get()
-            assertDoesNotThrow(() -> ConfigurationIO.as(YAML).writeAsync(instance).get());
+            assertDoesNotThrow(() -> instance.writeAsync(YAML));
 
             // Now, it should exists
-            assertTrue(ConfigurationIO.as(YAML).exist(instance));
+            assertTrue(instance.exist());
 
         }
         // Obtaining deleting task, then execute it
-        assertDoesNotThrow(() -> ConfigurationIO.as(YAML).deleteAsync(instance).get());
+        assertDoesNotThrow(() -> instance.deleteAsync().get());
         // Asserting does not exists any more
-        assertFalse(ConfigurationIO.as(YAML).exist(instance));
+        assertFalse(instance.exist());
 
     }