+++ CHANGELOG.md
@@ -8,7 +8,7 @@ depending on the time I have available to work on this library_
 
 ___
 
-## 0.0.3
+## 0.0.4
 
 **Release date:** [ *???* ]
 
@@ -19,11 +19,33 @@ ___
 - Nothing changed
 
 **Removed**
-- Nothing removed
+- Nothing changed
+___
+
+## 0.0.3
+
+**Release date:** [ *21 / 05 / 2020* ]
+
+**New**
+- Nothing changed
+
+**Changed**
+- Update README.md
+- Simplified I/O methods
+- ExportType renamed FormatType
+- Fixed wrong exceptions type on read() and write() methods
+- Fixed wrong documentation
+
+**Removed**
+- ConfigurationIO class
+
+___
+**Important** - This update break backward compatibility
+___
 
 ## 0.0.2
 
-**Release date:** [ *?? / 05 / 2020* ]
+**Release date:** [ *18 / 05 / 2020* ]
 
 **New**